+++ ui/src/test/java/de/ipb_halle/lbac/material/biomaterial/TaxonomyServiceTest.java
@@ -102,22 +102,22 @@ public class TaxonomyServiceTest extends TestBase {
         Taxonomy wulstlinge = taxonomies.get(12);
         Assert.assertEquals(7, wulstlinge.getLevel().getId());
         Assert.assertEquals("Wulstlinge_de", wulstlinge.getFirstName());
-        Assert.assertEquals(6, wulstlinge.getId());
-        Assert.assertEquals(5, wulstlinge.getTaxHierachy().size());
-        Assert.assertEquals(5, wulstlinge.getTaxHierachy().get(0).getId());
-        Assert.assertEquals(4, wulstlinge.getTaxHierachy().get(1).getId());
-        Assert.assertEquals(3, wulstlinge.getTaxHierachy().get(2).getId());
-        Assert.assertEquals(2, wulstlinge.getTaxHierachy().get(3).getId());
-        Assert.assertEquals(1, wulstlinge.getTaxHierachy().get(4).getId());
+        Assert.assertEquals(6,(int) wulstlinge.getId());
+        Assert.assertEquals(5,(int) wulstlinge.getTaxHierachy().size());
+        Assert.assertEquals(5,(int) wulstlinge.getTaxHierachy().get(0).getId());
+        Assert.assertEquals(4,(int) wulstlinge.getTaxHierachy().get(1).getId());
+        Assert.assertEquals(3,(int) wulstlinge.getTaxHierachy().get(2).getId());
+        Assert.assertEquals(2,(int) wulstlinge.getTaxHierachy().get(3).getId());
+        Assert.assertEquals(1,(int) wulstlinge.getTaxHierachy().get(4).getId());
 
         Taxonomy ohrlappenpilze = taxonomies.get(10);
         Assert.assertEquals(6, ohrlappenpilze.getLevel().getId());
         Assert.assertEquals("Gallerttränenverwandte_de", ohrlappenpilze.getFirstName());
-        Assert.assertEquals(11, ohrlappenpilze.getId());
-        Assert.assertEquals(3, ohrlappenpilze.getTaxHierachy().size());
-        Assert.assertEquals(8, ohrlappenpilze.getTaxHierachy().get(0).getId());
-        Assert.assertEquals(2, ohrlappenpilze.getTaxHierachy().get(1).getId());
-        Assert.assertEquals(1, ohrlappenpilze.getTaxHierachy().get(2).getId());
+        Assert.assertEquals(11,(int) ohrlappenpilze.getId());
+        Assert.assertEquals(3,(int) ohrlappenpilze.getTaxHierachy().size());
+        Assert.assertEquals(8,(int) ohrlappenpilze.getTaxHierachy().get(0).getId());
+        Assert.assertEquals(2,(int) ohrlappenpilze.getTaxHierachy().get(1).getId());
+        Assert.assertEquals(1,(int) ohrlappenpilze.getTaxHierachy().get(2).getId());
     }
 
     @Test
@@ -149,8 +149,8 @@ public class TaxonomyServiceTest extends TestBase {
         Assert.assertEquals("Haarnixen_de_edited", t.getNames().get(0).getValue());
         Assert.assertEquals("english_name", t.getNames().get(1).getValue());
         Assert.assertEquals(2, t.getTaxHierachy().size());
-        Assert.assertEquals(16, t.getTaxHierachy().get(0).getId());
-        Assert.assertEquals(1, t.getTaxHierachy().get(1).getId());
+        Assert.assertEquals(16, (int)t.getTaxHierachy().get(0).getId());
+        Assert.assertEquals(1,(int) t.getTaxHierachy().get(1).getId());
 
         Assert.assertEquals(3, t.getLevel().getId());
 
@@ -215,30 +215,30 @@ public class TaxonomyServiceTest extends TestBase {
         materialService.saveEditedMaterial(t3_copy, t3, project.getACList().getId(), owner.getId());
 
         t3 = service.loadTaxonomyById(3);
-        Assert.assertEquals(2, t3.getTaxHierachy().size());
-        Assert.assertEquals(2, t3.getTaxHierachy().get(0).getId());
-        Assert.assertEquals(0, t3.getTaxHierachy().get(1).getId());
-        Assert.assertEquals(14, t3.getLevel().getId());
+        Assert.assertEquals(2,(int) t3.getTaxHierachy().size());
+        Assert.assertEquals(2, (int)t3.getTaxHierachy().get(0).getId());
+        Assert.assertEquals(0,(int) t3.getTaxHierachy().get(1).getId());
+        Assert.assertEquals(14,(int) t3.getLevel().getId());
 
         Taxonomy t4 = service.loadTaxonomyById(4);
-        Assert.assertEquals(3, t4.getTaxHierachy().size());
-        Assert.assertEquals(3, t4.getTaxHierachy().get(0).getId());
-        Assert.assertEquals(2, t4.getTaxHierachy().get(1).getId());
-        Assert.assertEquals(0, t4.getTaxHierachy().get(2).getId());
-        Assert.assertEquals(18, t4.getLevel().getId());
+        Assert.assertEquals(3,(int) t4.getTaxHierachy().size());
+        Assert.assertEquals(3,(int) t4.getTaxHierachy().get(0).getId());
+        Assert.assertEquals(2,(int) t4.getTaxHierachy().get(1).getId());
+        Assert.assertEquals(0,(int) t4.getTaxHierachy().get(2).getId());
+        Assert.assertEquals(18,(int) t4.getLevel().getId());
         
         Taxonomy t5 = service.loadTaxonomyById(5);
-        Assert.assertEquals(3, t5.getTaxHierachy().size());
-        Assert.assertEquals(3, t5.getTaxHierachy().get(0).getId());
-        Assert.assertEquals(2, t5.getTaxHierachy().get(1).getId());
-        Assert.assertEquals(0, t5.getTaxHierachy().get(2).getId());
-        Assert.assertEquals(18, t5.getLevel().getId());
+        Assert.assertEquals(3,(int) t5.getTaxHierachy().size());
+        Assert.assertEquals(3,(int) t5.getTaxHierachy().get(0).getId());
+        Assert.assertEquals(2,(int) t5.getTaxHierachy().get(1).getId());
+        Assert.assertEquals(0,(int) t5.getTaxHierachy().get(2).getId());
+        Assert.assertEquals(18,(int) t5.getLevel().getId());
         
         Taxonomy t6 = service.loadTaxonomyById(6);
-        Assert.assertEquals(3, t6.getTaxHierachy().size());
-        Assert.assertEquals(3, t6.getTaxHierachy().get(0).getId());
-        Assert.assertEquals(2, t6.getTaxHierachy().get(1).getId());
-        Assert.assertEquals(0, t6.getTaxHierachy().get(2).getId());
+        Assert.assertEquals(3,(int) t6.getTaxHierachy().size());
+        Assert.assertEquals(3,(int) t6.getTaxHierachy().get(0).getId());
+        Assert.assertEquals(2,(int) t6.getTaxHierachy().get(1).getId());
+        Assert.assertEquals(0,(int) t6.getTaxHierachy().get(2).getId());
         Assert.assertEquals(18, t6.getLevel().getId());
     }
